pyproject.toml
[project]
name = "elephandroid"
version = "0.1.0"
description = "Microsoft 365 integrated enterprise automation tool"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "msal>=1.31.0",
    "httpx>=0.27.0",
    "openpyxl>=3.1.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.12",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.0",
    "python-pptx>=0.6.23",
    "python-docx>=1.1.0",
    "pymupdf>=1.24.0",
    "qdrant-client>=1.9.0",
    "alembic>=1.18.0",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.30.0"]
claude = ["anthropic>=0.40.0"]
openai = ["openai>=1.50.0"]
gemini = ["google-genai>=1.0.0"]
acs = ["azure-communication-callautomation>=1.4.0"]
all = ["asyncpg>=0.30.0", "anthropic>=0.40.0", "openai>=1.50.0", "google-genai>=1.0.0", "azure-communication-callautomation>=1.4.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
